fix(net-sockets): close CTS/socket leaks, harden teardown, expand tests
Annium.Net.Sockets review pass. Fix lifecycle/disposal leaks (terminal
_connectionCts and never-connected _listenCts; nativeSocket on failed
connect), a dropped reconnect cancellation token, the WhenConnected/
WaitForDisconnect handler leak on cancellation, the connect/disconnect
ObjectDisposedException races, and ConnectAsync returning the success
sentinel (null) on cancellation instead of OperationCanceledException.

Extract IConnectionMonitor (IConnectionMonitorFactory.Create now returns
it); add IServerSocket.IsConnected; unify ConnectionMonitorOptions
durations to int; create event TaskCompletionSources with
RunContinuationsAsynchronously.

Add 39 tests (91 -> 130) covering the above plus framing, disposal,
idempotency and cancellation paths; de-flake Listen_ConnectionLost,
Listen_Reconnect and the server-shutdown timing test; document new
public surface to keep docs-lint green.

+++ base/Net/src/Annium.Net.Sockets/ConnectionMonitorOptions.cs
@@ -18,5 +18,5 @@ public record ConnectionMonitorOptions
     /// <summary>
     /// Gets or sets the maximum delay allowed for ping responses in milliseconds (default: 300000ms).
     /// </summary>
-    public long MaxPingDelay { get; init; } = 300_000;
+    public int MaxPingDelay { get; init; } = 300_000;
 }
